Add classes to iconButton. Blocks removable.
Improved reusability of iconButton by allowing for classes to be passed down to iconButton to manage individual styling. Also allowed for passing down the img as a prop. Setup asystem for removing blocks from state. Had a number of problems with this including using index during .map and having all those indeces change during rerender. Also didn't bind this in the removeBlock function. Need to remember to not mutate state but always supply new state.