Redux throughout app. Working tree state structure. Update styling
The main improvements in this commit are that Redux is being used throughout the app, previously there were slices that weren't being used.The tree structure for the app's state is working. The tree state is housed in the sectionSlice which updates and maintains the tree state until it's saved which then saves it to the json file in /data/tree.json. TreeOperations.js does most of the heavy lifting for the tree traversal which is done through breadth-first-search. The tree, new node, and/or parent and current ids are passed to the various TreeOperation functions and they return the values needed to update the redux state. New icons and styling were also included in this commit as well as a connection point in preload that allows access to the main thread through BridgeContext.