Undefined var bug, loading initial state bug
Fix two bugs and updated a few styles. First bug was when choosing a variable on VariableModal it would show undefined as the next part of the sentence if the text ended in a variable expression. The conditional statement was evaluating loopIdx !== valueArr.length instead of valueArr.length-1. Fixed a small bug when the app is built where the initial json state needs loaded into userData folder for the first time. Handled this through a try catch block which will need updated since it's improper.