Refactor TreeOperations
Functions add and remove in TreeOperations used a common helper function _search to do their recursion. _search was originally hardcoded into each function for simplicity, but I wanted to remove the repeated code so I broke _search out into its own function and passed in the original arguments plus a callback so that it works for both functions.